# app/samples/data/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "HSKEE", "version": "0.1.0" },
  "paths": {
...
    "/contract/": {
      "get": {
        "tags": ["Contract"],
        "summary": "Get All",
        "operationId": "get_all_contract__get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DAOResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Contract"],
        "summary": "Create",
        "operationId": "create_contract__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContractCreateSchema" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DAOResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/contract/{id}": {
      "get": {
        "tags": ["Contract"],
        "summary": "Get",
        "operationId": "get_contract__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "uuid" },
                { "type": "string" }
              ],
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DAOResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Contract"],
        "summary": "Update",
        "operationId": "update_contract__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContractUpdateSchema" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DAOResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Contract"],
        "summary": "Delete",
        "operationId": "delete_contract__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Address": {
        "properties": {
          "address_type": { "$ref": "#/components/schemas/AddressTypeEnum" },
          "primary": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Primary",
            "default": true
          },
          "address_1": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address 1"
          },
          "address_2": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address 2"
          },
          "city": {
            "anyOf": [
              { "type": "string" },
              { "type": "string", "format": "uuid" },
              {}
            ],
            "title": "City"
          },
          "region": {
            "anyOf": [
              { "type": "string" },
              { "type": "string", "format": "uuid" },
              {}
            ],
            "title": "Region"
          },
          "country": {
            "anyOf": [
              { "type": "string" },
              { "type": "string", "format": "uuid" },
              {}
            ],
            "title": "Country"
          },
          "address_postalcode": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address Postalcode"
          },
          "address_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Address Id"
          }
        },
        "type": "object",
        "required": [
          "address_type",
          "address_1",
          "city",
          "region",
          "country",
          "address_postalcode"
        ],
        "title": "Address",
        "description": "Model for representing an address with an optional address ID.\n\nAttributes:\n    address_id (Optional[UUID]): The unique identifier for the address, defaults to None."
      },
...
      "AddressTypeEnum": {
        "type": "string",
        "enum": ["billing", "mailing"],
        "title": "AddressTypeEnum",
        "description": "Enumeration for address types.\n\nAttributes:\n    billing (str): Represents a billing address.\n    mailing (str): Represents a mailing address."
      },
...
      "ContractCreateSchema": {
        "properties": {
          "contract_type": { "type": "string", "title": "Contract Type" },
          "payment_type": { "type": "string", "title": "Payment Type" },
          "contract_status": { "$ref": "#/components/schemas/ContractStatus" },
          "contract_details": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Contract Details"
          },
          "payment_amount": {
            "anyOf": [{ "type": "number" }, { "type": "string" }],
            "title": "Payment Amount"
          },
          "fee_percentage": {
            "anyOf": [
              { "type": "number" },
              { "type": "string" },
              { "type": "null" }
            ],
            "title": "Fee Percentage"
          },
          "fee_amount": {
            "anyOf": [
              { "type": "number" },
              { "type": "string" },
              { "type": "null" }
            ],
            "title": "Fee Amount"
          },
          "date_signed": {
            "type": "string",
            "format": "date-time",
            "title": "Date Signed",
            "default": "2024-07-21T12:23:52.479870"
          },
          "start_date": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Start Date",
            "default": "2024-07-21T12:23:52.479877"
          },
          "end_date": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "End Date"
          },
          "contract_info": {
            "anyOf": [
              {
                "items": { "$ref": "#/components/schemas/UnderContractSchema" },
                "type": "array"
              },
              { "$ref": "#/components/schemas/UnderContractSchema" },
              { "type": "null" }
            ],
            "title": "Contract Info"
          },
          "utilities": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EntityBillableCreate"
                },
                "type": "array"
              },
              { "$ref": "#/components/schemas/EntityBillableCreate" },
              { "type": "null" }
            ],
            "title": "Utilities"
          }
        },
        "type": "object",
        "required": [
          "contract_type",
          "payment_type",
          "contract_status",
          "payment_amount"
        ],
        "title": "ContractCreateSchema",
        "description": "Schema for creating a new contract.\n\nAttributes:\n    contract_type (str): The type of the contract.\n    payment_type (str): The type of payment for the contract.\n    contract_status (ContractStatus): The status of the contract.\n    contract_details (Optional[str]): The details of the contract.\n    payment_amount (Decimal): The payment amount for the contract.\n    fee_percentage (Optional[Decimal]): The fee percentage for the contract.\n    fee_amount (Optional[Decimal]): The fee amount for the contract.\n    date_signed (datetime): The date the contract was signed.\n    start_date (Optional[datetime]): The start date of the contract.\n    end_date (Optional[datetime]): The end date of the contract.\n    contract_info (Optional[List[UnderContractSchema] | UnderContractSchema]): The information of the under-contract relationships.\n    utilities (Optional[List[EntityBillableCreate] | EntityBillableCreate]): The utilities associated with the contract."
      },
      "ContractStatus": {
        "type": "string",
        "enum": ["active", "expired", "terminated"],
        "title": "ContractStatus",
        "description": "Enumeration for contract status types.\n\nAttributes:\n    active (str): Represents an active contract.\n    expired (str): Represents an expired contract.\n    terminated (str): Represents a terminated contract."
      },
...
      "ContractUpdateSchema": {
        "properties": {
          "contract_type": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Contract Type"
          },
          "payment_type": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Payment Type"
          },
          "contract_status": {
            "anyOf": [
              { "$ref": "#/components/schemas/ContractStatus" },
              { "type": "null" }
            ]
          },
          "contract_details": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Contract Details"
          },
          "payment_amount": {
            "anyOf": [
              { "type": "number" },
              { "type": "string" },
              { "type": "null" }
            ],
            "title": "Payment Amount"
          },
          "fee_percentage": {
            "anyOf": [
              { "type": "number" },
              { "type": "string" },
              { "type": "null" }
            ],
            "title": "Fee Percentage"
          },
          "fee_amount": {
            "anyOf": [
              { "type": "number" },
              { "type": "string" },
              { "type": "null" }
            ],
            "title": "Fee Amount"
          },
          "date_signed": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Date Signed",
            "default": "2024-07-21T12:23:52.539497"
          },
          "start_date": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Start Date",
            "default": "2024-07-21T12:23:52.539504"
          },
          "end_date": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "End Date"
          },
          "contract_info": {
            "anyOf": [
              {
                "items": { "$ref": "#/components/schemas/UnderContractSchema" },
                "type": "array"
              },
              { "$ref": "#/components/schemas/UnderContractSchema" },
              { "type": "null" }
            ],
            "title": "Contract Info"
          },
          "utilities": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EntityBillableCreate"
                },
                "type": "array"
              },
              { "$ref": "#/components/schemas/EntityBillableCreate" },
              { "type": "null" }
            ],
            "title": "Utilities"
          }
        },
        "type": "object",
        "title": "ContractUpdateSchema",
        "description": "Schema for updating an existing contract.\n\nAttributes:\n    contract_type (Optional[str]): The type of the contract.\n    payment_type (Optional[str]): The type of payment for the contract.\n    contract_status (Optional[ContractStatus]): The status of the contract.\n    contract_details (Optional[str]): The details of the contract.\n    payment_amount (Optional[Decimal]): The payment amount for the contract.\n    fee_percentage (Optional[Decimal]): The fee percentage for the contract.\n    fee_amount (Optional[Decimal]): The fee amount for the contract.\n    date_signed (Optional[datetime]): The date the contract was signed.\n    start_date (Optional[datetime]): The start date of the contract.\n    end_date (Optional[datetime]): The end date of the contract.\n    contract_info (Optional[List[UnderContractSchema] | UnderContractSchema]): The information of the under-contract relationships.\n    utilities (Optional[List[EntityBillableCreate] | EntityBillableCreate]): The utilities associated with the contract."
      },
      "DAOResponse": { "type": "object", "title": "DAOResponse" },
...
      "EntityBillableCreate": {
        "properties": {
          "billable_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Billable Id"
          },
          "payment_type": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Payment Type"
          },
          "billable_amount": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Billable Amount"
          },
          "apply_to_units": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Apply To Units",
            "default": false
          }
        },
        "type": "object",
        "title": "EntityBillableCreate",
        "description": "Schema for creating an entity billable.\n\nAttributes:\n    billable_id (Optional[UUID]): The unique identifier for the billable entity.\n    payment_type (Optional[str]): The type of payment for the billable entity.\n    billable_amount (Optional[int]): The amount for the billable entity.\n    apply_to_units (Optional[bool]): Indicates whether the billable entity applies to units."
      },
...
      "GenderEnum": {
        "type": "string",
        "enum": ["male", "female", "other"],
        "title": "GenderEnum",
        "description": "Enumeration for gender types.\n\nAttributes:\n    male (str): Represents male gender.\n    female (str): Represents female gender.\n    other (str): Represents other genders."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Property": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "property_type": { "$ref": "#/components/schemas/PropertyType" },
          "amount": { "type": "number", "title": "Amount" },
          "security_deposit": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Security Deposit"
          },
          "commission": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Commission"
          },
          "floor_space": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Floor Space"
          },
          "num_units": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Num Units"
          },
          "num_bathrooms": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Num Bathrooms"
          },
          "num_garages": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Num Garages"
          },
          "has_balconies": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Has Balconies",
            "default": false
          },
          "has_parking_space": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Has Parking Space",
            "default": false
          },
          "pets_allowed": {
            "type": "boolean",
            "title": "Pets Allowed",
            "default": false
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "property_status": { "$ref": "#/components/schemas/PropertyStatus" },
          "property_unit_assoc_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Property Unit Assoc Id"
          },
          "address": {
            "anyOf": [
              {
                "items": { "$ref": "#/components/schemas/Address" },
                "type": "array"
              },
              { "$ref": "#/components/schemas/Address" },
              { "type": "null" }
            ],
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "name",
          "property_type",
          "amount",
          "property_status",
          "property_unit_assoc_id"
        ],
        "title": "Property",
        "description": "Model for representing a property with additional details.\n\nAttributes:\n    property_unit_assoc_id (Optional[UUID]): The unique identifier for the property unit association.\n    address (Optional[List[Address] | Address]): The address(es) associated with the property."
      },
...
      "PropertyStatus": {
        "type": "string",
        "enum": ["available", "unavailable"],
        "title": "PropertyStatus",
        "description": "Enumeration for property status types.\n\nAttributes:\n    available (str): Represents an available property status.\n    unavailable (str): Represents an unavailable property status."
      },
      "PropertyType": {
        "type": "string",
        "enum": ["residential", "commercial", "industrial"],
        "title": "PropertyType",
        "description": "Enumeration for property types.\n\nAttributes:\n    residential (str): Represents a residential property.\n    commercial (str): Represents a commercial property.\n    industrial (str): Represents an industrial property."
      },
      "PropertyUnit": {
        "properties": {
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "property_unit_code": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Property Unit Code"
          },
          "property_unit_floor_space": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Property Unit Floor Space"
          },
          "property_unit_amount": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Property Unit Amount"
          },
          "property_floor_id": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Property Floor Id"
          },
          "property_unit_notes": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Property Unit Notes"
          },
          "property_unit_security_deposit": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Property Unit Security Deposit"
          },
          "property_unit_commission": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Property Unit Commission"
          },
          "has_amenities": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Has Amenities",
            "default": false
          },
          "property_unit_assoc_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Property Unit Assoc Id"
          }
        },
        "type": "object",
        "required": ["property_id", "property_unit_assoc_id"],
        "title": "PropertyUnit",
        "description": "Model for representing a property unit with additional details.\n\nAttributes:\n    property_unit_assoc_id (Optional[UUID]): The unique identifier for the property unit association."
      },
...
      "UnderContractSchema": {
        "properties": {
          "under_contract_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Under Contract Id"
          },
          "property_unit_assoc": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "$ref": "#/components/schemas/Property" },
              { "$ref": "#/components/schemas/PropertyUnit" },
              { "type": "null" }
            ],
            "title": "Property Unit Assoc"
          },
          "contract_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Contract Id"
          },
          "contract_status": {
            "anyOf": [
              { "$ref": "#/components/schemas/ContractStatus" },
              { "type": "null" }
            ]
          },
          "client_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "$ref": "#/components/schemas/UserBase" },
              { "type": "null" }
            ],
            "title": "Client Id"
          },
          "employee_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "$ref": "#/components/schemas/UserBase" },
              { "type": "null" }
            ],
            "title": "Employee Id"
          }
        },
        "type": "object",
        "title": "UnderContractSchema",
        "description": "Schema for representing an under-contract relationship.\n\nAttributes:\n    under_contract_id (Optional[UUID]): The unique identifier for the under-contract relationship.\n    property_unit_assoc (Optional[UUID | Property | PropertyUnit]): The associated property or property unit.\n    contract_id (Optional[UUID]): The unique identifier for the contract.\n    contract_status (Optional[ContractStatus]): The status of the contract.\n    client_id (Optional[UUID | UserBase]): The unique identifier for the client.\n    employee_id (Optional[UUID | UserBase]): The unique identifier for the employee."
      },
...
      "UserBase": {
        "properties": {
          "date_of_birth": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "null" }
            ],
            "title": "Date Of Birth"
          },
          "first_name": { "type": "string", "title": "First Name" },
          "last_name": { "type": "string", "title": "Last Name" },
          "email": { "type": "string", "format": "email", "title": "Email" },
          "phone_number": { "type": "string", "title": "Phone Number" },
          "identification_number": {
            "type": "string",
            "title": "Identification Number"
          },
          "photo_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Photo Url",
            "default": ""
          },
          "gender": { "$ref": "#/components/schemas/GenderEnum" }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "email",
          "phone_number",
          "identification_number",
          "gender"
        ],
        "title": "UserBase",
        "description": "Base model for user information.\n\nAttributes:\n    date_of_birth (Optional[date]): The date of birth of the user.\n    first_name (str): The first name of the user (max length 128).\n    last_name (str): The last name of the user (max length 128).\n    email (EmailStr): The email address of the user.\n    phone_number (str): The phone number of the user (max length 50).\n    identification_number (str): The identification number of the user (max length 80).\n    photo_url (Optional[str]): The URL of the user's photo.\n    gender (GenderEnum): The gender of the user."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}